# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Security scanning platform — manages targets, scan jobs, reports, and triage rules.",
        "title": "NexusSec API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Xác thực user. Nếu thành công, gửi OTP qua email để xác thực 2 bước.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Thông tin đăng nhập",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_handler.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP đã gửi về email",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Sai email hoặc mật khẩu",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Tài khoản chưa xác thực email",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy thông tin profile của user đang đăng nhập (yêu cầu JWT).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gateway_handler.userResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/auth/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Đổi mật khẩu. Yêu cầu JWT và mật khẩu hiện tại đúng.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Mật khẩu cũ và mới",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_handler.changePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đổi mật khẩu thành công",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Mật khẩu hiện tại sai",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Tạo tài khoản mới. Password được hash bằng bcrypt (cost=12). Trả về 409 nếu email/username đã tồn tại.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Thông tin đăng ký",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_handler.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gateway_handler.userResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "409": {
                        "description": "Email hoặc username đã tồn tại",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_nexussec_nexussec_pkg_response.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_gateway_handler.changePasswordRequest": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                }
            }
        },
...
        "internal_gateway_handler.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_gateway_handler.registerRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
...
        "internal_gateway_handler.userResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Nhập: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}